# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Techpack Solutions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversationMessageRole = typeof ConversationMessageRole[keyof typeof ConversationMessageRole];


export const ConversationMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ConversationMessage {
  role: ConversationMessageRole;
  content: string;
}

export interface ChatMessageInput {
  message: string;
  conversationHistory: ConversationMessage[];
}

export interface ChatResponse {
  reply: string;
}

export interface LeadInput {
  name: string;
  phone: string;
  location: string;
  /** @nullable */
  notes?: string | null;
}

export interface Lead {
  id: number;
  name: string;
  phone: string;
  location: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tag?: string | null;
  isActive?: boolean;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tag?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  location: string;
  productName: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  new: 'new',
  contacted: 'contacted',
  quoted: 'quoted',
  closed: 'closed',
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  location: string;
  productName: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  notes?: string | null;
  status: OrderStatus;
  createdAt: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  new: 'new',
  contacted: 'contacted',
  quoted: 'quoted',
  closed: 'closed',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}