# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Techpack Solutions API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: chat
  - name: leads
  - name: products
  - name: orders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /leads:
    post:
      operationId: captureLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    get:
      operationId: getLeads
      tags: [leads]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all active products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ChatMessageInput:
      type: object
      required: [message, conversationHistory]
      properties:
        message:
          type: string
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"

    ConversationMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    ChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    LeadInput:
      type: object
      required: [name, phone, location]
      properties:
        name:
          type: string
        phone:
          type: string
        location:
          type: string
        notes:
          type: ["string", "null"]

    Lead:
      type: object
      required: [id, name, phone, location, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        location:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    ProductInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        tag:
          type: ["string", "null"]
        isActive:
          type: boolean

    Product:
      type: object
      required: [id, name, description, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        tag:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time

    OrderInput:
      type: object
      required: [customerName, phone, location, productName]
      properties:
        customerName:
          type: string
        phone:
          type: string
        location:
          type: string
        productName:
          type: string
        quantity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    Order:
      type: object
      required: [id, customerName, phone, location, productName, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        location:
          type: string
        productName:
          type: string
        quantity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, quoted, closed]
        createdAt:
          type: string
          format: date-time

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, contacted, quoted, closed]