# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Techpack Solutions API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessage } from './conversationMessage';

export interface ChatMessageInput {
  message: string;
  conversationHistory: ConversationMessage[];
}