# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Techpack Solutions API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  location: string;
  productName: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  notes?: string | null;
  status: OrderStatus;
  createdAt: Date;
}