# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Techpack Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  customerName: string;
  phone: string;
  location: string;
  productName: string;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  notes?: string | null;
}