# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Techpack Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  tag?: string | null;
  isActive: boolean;
  createdAt: Date;
}